requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.2
urllib3>=2.1.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.1
numpy>=1.24.3
scipy>=1.11.4
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
loguru>=0.7.2
nltk>=3.8.1
spacy>=3.4.0
tqdm>=4.66.1
PyYAML>=6.0.1
